refactor(tags): extract name and tag modules
